routes and refact
