+++ helpers/resources.js
@@ -2,12 +2,13 @@
 
 const express = require('express');
 const router  = express.Router();
-const helpers = require('../helpers/index');
 
 module.exports = (knex) => {
+  const helpers = require('../helpers/index')(knex);
 
-  deleteResource: async function(req, res, token, knex){
-    const userId = await helpers.getUserId(token);
+  return{
+    deleteResource: async function(req, res, token, knex){
+    const userId = await helpers.getUserIdByToken(token);
     console.log(userId);
     // check if it is the user who created it
     knex
@@ -24,10 +25,10 @@ module.exports = (knex) => {
           .then(res.status(200).json( {success: 'Deleted' }))
       })
       .catch(e => res.status(400).json( {e} ));
-  }
+  },
 
   updateResource: async function  (req, res, token, knex){
-    const userId = await helpers.getUserId(token);
+    const userId = await helpers.getUserIdByToken(token);
     console.log(userId);
     const updateResource = {};
     const updateRefrences = {};
@@ -85,10 +86,10 @@ module.exports = (knex) => {
 
               const length = results.length;
               const referenceFound = results.filter(result => result.user_id === userId);
+              const maxId = Math.max.apply(Math, results.map(result => result.id ));
 
               // create new refrence
               if(length === 0 || referenceFound.length === 0){
-                const maxId = Math.max.apply(Math, results.map(result => result.id ));
                 const createReference = {
                     id: maxId++,
                     resource_id: req.params.id,
@@ -124,25 +125,26 @@ module.exports = (knex) => {
                       updtReference.user_id = userId;
                       return res.status(200).json( updtReference );
                     }
+                    res.status(400).json( {error: "Couldn't create resource references"} )
                   })
               }
             })
         }
       })
       .catch(e => res.status(400).json( {e} ));
-  }
+  },
 
   createResource: async function  (req, res, token, knex){
-    const userId = await helpers.getUserId(token);
+    const userId = await helpers.getUserIdByToken(token);
     console.log(userId);
 
     knex('resources').max('id')
       .then(result => result[0].max + 1)
       .then( max => {
 
-        const { url, title, description, category_id, } = req.body;
+        const { url, title, description, category_id } = req.body;
         const newResource = {
-          id: max++,
+          id: max,
           url,
           title,
           description,
@@ -156,13 +158,103 @@ module.exports = (knex) => {
         }
 
         knex('resources').insert({newResource})
-          .then(results =>  res.json(newResource));
+          .then(results =>  results.length === 1 ? res.json(newResource) : res.status(400).json( {error: "Couldn't create resource"} ));
       })
+      .catch(e => res.status(400).json( e ));
+  },
+
+  getReources: (req, res) => {
+
+    knex
+      .select('res.id', 'res.url', 'res.title', 'res.description', 'res.created_on',
+        'res.created_by as created_by_id', 'users.username',
+        knex.raw(`sum(rate)/count(ref.user_id) as rate`),
+        knex.raw(`sum(case
+            when liked = true then 1
+            else 0
+            end) as likes`),
+        knex.raw('count(com.id)  as comments'),
+        'cat.id as category_id', 'cat.description as category_description')
+      .from("resources as res")
+      .innerJoin('users', 'users.id', 'res.created_by')
+      .innerJoin('resources_references as ref', 'ref.resource_id', 'res.id')
+      .innerJoin('rates', 'rates.id', 'ref.rate_id ')
+      .innerJoin('comments as com', 'com.resource_id', 'res.id ')
+      .innerJoin('categories as cat', 'cat.id', 'res.category_id ')
+      .groupBy('res.id', 'users.username', 'cat.id')
+      .where('res.id', req.params.id)
+      .then(results => res.json(results))
       .catch(e => res.status(400).json( {e} ));
+  },
+
+  getReourcesByQuery: (req, res) => {
+
+    const parsedUrl = url.parse(req.originalUrl);
+    const parsedQs = querystring.parse(parsedUrl.query);
+    const { search, limit } = parsedQs;
+
+    if(search){
+
+      const searchFor = `%${ search }%`;
+
+      knex
+        .select('res.id', 'res.url', 'res.title', 'res.description', 'res.created_on',
+          'res.created_by as created_by_id', 'users.username',
+          knex.raw(`sum(rate)/count(ref.user_id) as rate`),
+          knex.raw(`sum(case
+              when liked = true then 1
+              else 0
+              end) as likes`),
+          knex.raw('count(com.id)  as comments'),
+          'cat.id as category_id', 'cat.description as category_description')
+        .from("resources as res")
+        .innerJoin('users', 'users.id', 'res.created_by')
+        .innerJoin('resources_references as ref', 'ref.resource_id', 'res.id')
+        .innerJoin('rates', 'rates.id', 'ref.rate_id ')
+        .innerJoin('comments as com', 'com.resource_id', 'res.id ')
+        .innerJoin('categories as cat', 'cat.id', 'res.category_id ')
+        .where('cat.description', 'ilike', searchFor)
+        .orWhere('res.url', 'ilike', searchFor)
+        .orWhere('res.title', 'ilike', searchFor)
+        .orWhere('res.description', 'ilike', searchFor)
+        .groupBy('res.id', 'users.username', 'cat.id')
+        .orderBy('rate', 'desc')
+        .limit(limit)
+        .then( results => res.json(results))
+        .catch(e => res.status(400).json( {e} ));
+
+    }else{
+
+      if(limit){
+        knex
+          .select('res.id', 'res.url', 'res.title', 'res.description', 'res.created_on',
+            'res.created_by as created_by_id', 'users.username',
+            knex.raw(`sum(rate)/count(ref.user_id) as rate`),
+            knex.raw(`sum(case
+                when liked = true then 1
+                else 0
+                end) as likes`),
+            knex.raw('count(com.id)  as comments'),
+            'cat.id as category_id', 'cat.description as category_description')
+          .from("resources as res")
+          .innerJoin('users', 'users.id', 'res.created_by')
+          .innerJoin('resources_references as ref', 'ref.resource_id', 'res.id')
+          .innerJoin('rates', 'rates.id', 'ref.rate_id ')
+          .innerJoin('comments as com', 'com.resource_id', 'res.id ')
+          .innerJoin('categories as cat', 'cat.id', 'res.category_id ')
+          .groupBy('res.id', 'users.username', 'cat.id')
+          .orderBy('rate', 'desc')
+          .limit(limit)
+          .then( results => res.json(results))
+          .catch(e => res.status(400).json( {e} ));
+      }
+    }
   }
+  }
+
 }
 // async function deleteResource (req, res, token, knex){
-//   const userId = await helpers.getUserId(token);
+//   const userId = await helpers.getUserIdByToken(token);
 
 //   // check if it is the user who created it
 //   knex
@@ -183,7 +275,7 @@ module.exports = (knex) => {
 
 // async function updateResource (req, res, token, knex){
 
-//   const userId = await helpers.getUserId(token);
+//   const userId = await helpers.getUserIdByToken(token);
 //   const updateResource = {};
 //   const updateRefrences = {};
